+++ randfractal-jmelange/src/main/java/com/github/terefang/jmelange/randfractal/lite/FnSolid.java
@@ -27,7 +27,7 @@ public class FnSolid extends FastNoiseLite
     static int[] TABLE3 = null;
 
     // 2d
-    public static final float singleSolid(int interpolation, int seed, float x, float y)
+    public static final double singleSolid(int interpolation, int seed, double x, double y)
     {
         mkTable();
 
@@ -36,7 +36,7 @@ public class FnSolid extends FastNoiseLite
         int x1 = x0 + 1;
         int y1 = y0 + 1;
 
-        float xs, ys;
+        double xs, ys;
         switch (interpolation) {
             default:
             case LINEAR:
@@ -53,29 +53,29 @@ public class FnSolid extends FastNoiseLite
                 break;
         }
 
-        float xd0 = x - x0;
-        float yd0 = y - y0;
-        float xd1 = xd0 - 1;
-        float yd1 = yd0 - 1;
+        double xd0 = x - x0;
+        double yd0 = y - y0;
+        double xd1 = xd0 - 1;
+        double yd1 = yd0 - 1;
 
-        float xf0 = lerp(table2D(seed, x0, y0, xd0, yd0), table2D(seed, x1, y0, xd1, yd0), xs);
-        float xf1 = lerp(table2D(seed, x0, y1, xd0, yd1), table2D(seed, x1, y1, xd1, yd1), xs);
+        double xf0 = lerp(table2D(seed, x0, y0, xd0, yd0), table2D(seed, x1, y0, xd1, yd0), xs);
+        double xf1 = lerp(table2D(seed, x0, y1, xd0, yd1), table2D(seed, x1, y1, xd1, yd1), xs);
 
         return lerp(xf0, xf1, ys);
     }
 
-    private static float table2D(int _seed, int _x0, int _y0, float _xd, float _yd)
+    private static double table2D(int _seed, int _x0, int _y0, double _xd, double _yd)
     {
         int _hash = hash256(_x0, _y0, _seed) & -2 ;
-        float _n0 = 128f-((float)TABLE1[_hash]);
-        float _n1 = 128f-((float)TABLE2[_hash]);
-        float _n2 = 128f-((float)TABLE1[_hash+1]);
-        float _n3 = 128f-((float)TABLE2[_hash+1]);
+        double _n0 = 128f-((double)TABLE1[_hash]);
+        double _n1 = 128f-((double)TABLE2[_hash]);
+        double _n2 = 128f-((double)TABLE1[_hash+1]);
+        double _n3 = 128f-((double)TABLE2[_hash+1]);
 
         return _xd*(_n0/128f)+_yd*(_n1/128f)-(_n2/255f)+(_n3/511f);
     }
 
-    public static final float singleSolid(int interpolation, int seed, float x, float y, float z)
+    public static final double singleSolid(int interpolation, int seed, double x, double y, double z)
     {
         mkTable();
 
@@ -86,7 +86,7 @@ public class FnSolid extends FastNoiseLite
         int y1 = y0 + 1;
         int z1 = z0 + 1;
 
-        float xs, ys, zs;
+        double xs, ys, zs;
         switch (interpolation) {
             default:
             case LINEAR:
@@ -106,33 +106,33 @@ public class FnSolid extends FastNoiseLite
                 break;
         }
 
-        final float xd0 = x - x0;
-        final float yd0 = y - y0;
-        final float zd0 = z - z0;
-        final float xd1 = xd0 - 1;
-        final float yd1 = yd0 - 1;
-        final float zd1 = zd0 - 1;
+        final double xd0 = x - x0;
+        final double yd0 = y - y0;
+        final double zd0 = z - z0;
+        final double xd1 = xd0 - 1;
+        final double yd1 = yd0 - 1;
+        final double zd1 = zd0 - 1;
 
-        final float xf00 = lerp(table3D(seed, x0, y0, z0, xd0, yd0, zd0), table3D(seed, x1, y0, z0, xd1, yd0, zd0), xs);
-        final float xf10 = lerp(table3D(seed, x0, y1, z0, xd0, yd1, zd0), table3D(seed, x1, y1, z0, xd1, yd1, zd0), xs);
-        final float xf01 = lerp(table3D(seed, x0, y0, z1, xd0, yd0, zd1), table3D(seed, x1, y0, z1, xd1, yd0, zd1), xs);
-        final float xf11 = lerp(table3D(seed, x0, y1, z1, xd0, yd1, zd1), table3D(seed, x1, y1, z1, xd1, yd1, zd1), xs);
+        final double xf00 = lerp(table3D(seed, x0, y0, z0, xd0, yd0, zd0), table3D(seed, x1, y0, z0, xd1, yd0, zd0), xs);
+        final double xf10 = lerp(table3D(seed, x0, y1, z0, xd0, yd1, zd0), table3D(seed, x1, y1, z0, xd1, yd1, zd0), xs);
+        final double xf01 = lerp(table3D(seed, x0, y0, z1, xd0, yd0, zd1), table3D(seed, x1, y0, z1, xd1, yd0, zd1), xs);
+        final double xf11 = lerp(table3D(seed, x0, y1, z1, xd0, yd1, zd1), table3D(seed, x1, y1, z1, xd1, yd1, zd1), xs);
 
-        final float yf0 = lerp(xf00, xf10, ys);
-        final float yf1 = lerp(xf01, xf11, ys);
+        final double yf0 = lerp(xf00, xf10, ys);
+        final double yf1 = lerp(xf01, xf11, ys);
 
         return lerp(yf0, yf1, zs);
     }
 
-    private static float table3D(int _seed, int _x0, int _y0, int _z0, float _xd, float _yd, float _zd)
+    private static double table3D(int _seed, int _x0, int _y0, int _z0, double _xd, double _yd, double _zd)
     {
         int _hash = hash256(_x0, _y0, _z0, _seed) & -2 ;
-        float _n0 = 128f-((float)TABLE1[_hash]);
-        float _n1 = 128f-((float)TABLE2[_hash]);
-        float _n2 = 128f-((float)TABLE3[_hash]);
-        float _n3 = 128f-((float)TABLE1[_hash+1]);
-        float _n4 = 128f-((float)TABLE2[_hash+1]);
-        float _n5 = 128f-((float)TABLE3[_hash+1]);
+        double _n0 = 128f-((double)TABLE1[_hash]);
+        double _n1 = 128f-((double)TABLE2[_hash]);
+        double _n2 = 128f-((double)TABLE3[_hash]);
+        double _n3 = 128f-((double)TABLE1[_hash+1]);
+        double _n4 = 128f-((double)TABLE2[_hash+1]);
+        double _n5 = 128f-((double)TABLE3[_hash+1]);
 
         return _xd*(_n0/128f)+_yd*(_n1/128f)+_zd*(_n2/128f)-(_n3/255f)+(_n4/511f)-(_n5/1023f);
     }
